requirements.txt
--index-url https://download.pytorch.org/whl/cpu
torch
fastapi
uvicorn
pydantic

langchain
langchain-community
langchain-groq

chromadb

sentence-transformers

rank-bm25

pandas
numpy

python-dotenv